Code introduced to remove ghosts with the charge correlation between the 2 cathods. A chi2 is performed for the 2 possibilities. If one gets good chi2 (with respect to the fGhostChi2Cut parameter) and the other wrong chi2, the ambiguity is solved If both gets good or both bad chi2, no choice is made By default the fGhostChi2Cut parameter is set to solve about 70% of ghost problems with about 2% errors, with the current version of the code.
Implementation :
fGhostChi2Cut is in AliMUONClusterFinderVS, with setters and getters.
a fDebugLevel was also introduced to switch off some of the output.
When an ambiguity is detected and not solved, the fGhost word in
AliMUONRawCluster is set to 1 or 2, depending whether both charge chi2 are
good or bad.
a DumpIndex method was also added in AliMUONRawCluster to produce a printout
of digit indexes.

User incidences :
By default, the code makes ghost check. If you want previous behaviour,
put in MUONrawclusters the value of SetGhostChi2Cut to infinity (1e.6) is
sufficient.