Some changes resulting from last week work. The mc stuff has been removed. Added possibility to not save the width when compressing file; this is one by setting the flag AliL3Compress::fWriteShape.
The status of the code is that it seems to be more or less working. The
efficiency drops a little bit when running it through offline, but this
is mainly because overlapping clusters are not deconvoluted, and 2 clusters
found by the offline cluster finder will be 1 cluster after compressing/
decompressing. The only way to get rid of this, is to include some
deconvolution in the AliL3Modeller::FindClusters.