Moved to the latest version of the HLT code in Bergen. Most important changes, since last checkin are:
a) Static Index in AliL3Filehandler to speed up IO

b) LUT Hough transformation improving track candidate
finding by a factor of 3

c) Changes related to how the track parameters are calculated and stored.
There should be no changes for normal use, i.e. the function calls are
still the same. The changes are the following:
- Most importantly, the first point on a track (accessible via
AliL3Track::GetFirstPoint()) is now calculated as the point lying on the helix
fit. Earlier, this point was just the coordinates of the innermost assigned
cluster of the track. This is in principle ok if you just want the quick
"global" coordinates of the track; momentum etc.  However, for precisce local
calculations such as residuals, this point should be the point of closest
approach lying on the helix fit. If not, the calculated track parameters (such
as center of curvature) will be slightly shifted. The tracks are updated to the
first point if the AliLevel3::fIsTracksAtFirstPoint is set to true (which is
default). If this is not desired, and the track should be defined at the vertex
(not DCAO!), you can still do AliLevel3::SetTracks2Vertex().

d) A new static boolean has been introduced in AliLevel3::fDoVertexFit, which
tells the track fitter to include the vertex in the fit or not.

e)Change in interface of updating tracks, instead of using the fitter class,
use track->UpdatetoFistPoint (see deconvoluter macro).