Got rid of class template AliFMD<Type> on request of Federico, who thinks they are evil (sigh!).   Added a lot of (duplicate) code to cover what was previously done by that class template.
New structure to the reconstruction algorithms.  In essence, the
details of the algorithms areput into separate classes.

Clean-up of some code.

Removed header guards in implementation files _only_ - Peter thinks
they are evil (sigh!).  They are still present in the declaration files
as I believe that is definitely the Right Thing (tm) to do cf. Fons.

The code is in a pretty unstable state, but I had to commit as Federico
and Peter wanted the class template out ASAP.  Apparently there's a
problem on Optreon/Itianium machines where either R__ACCESS_IN_SYMBOL
or R__USE_SHADOW_CLASS is defined, and an assert fails:

   Assert(sizeof(AliFMDMap<float>) == sizeof(::ROOT::Shadow::BlaBla))

This seems to be a problem in (ROOT)CINT - not with templates as such.