Commiting before moving code from `AliFMDSimulator' to `AliFMD' (and sub-classes), as `requested' by Federico.   The definition of materials, and the like will move to `AliFMD' directly - cluttering up the class even more than it is already (sigh!).    The step-management will go in `AliFMDv1' (and relevant sub-classes).  The geometry definition will go in the new class `AliFMDGeometryBuilder' (and sub-classes), to at least keep that out of the `AliFMD' and to make it easier to make changes to the geometry (via inheritance).
Federico also complains about the size of the digits.   I'm not sure that
I understand his complaint, but we'll see.

The `AliFMDG3Simulator' (and sub-classes) will disappear completely.  That
means, that the VMC _must_ support a geometry based on the ROOT `TGeo'
classes - rather than via calls to the GEANT 3.21-like member function
calls.   This commit will be tagged.

I've temporarily added the class `AliFMDAlla' which is Alla's old
implementation of the FMD geometry.   This class will disappear.