I've tried to make some output messages lighter (and moved others to the debug stream) according to Federico's requests.
I've also upgraded the access framework following a proposal by Latchezar.
I've added a protection in the Put methods that should prevent users to
put "anything" into OCDB folders (where only OCDB data should stay). The
user should now declare explicitly (through this flag) if he wants to
store an object in an OCDB folder, in a "Reference" folder or somewhere
else (in a "private" folder). If the data is flagged as "reference" or
"private" then it cannot be put in the (currently hard-coded in
AliCDBManager) official OCDB folder.

This new desing is still under discussion (for instance, currently the
OCDB and reference folder are two test folders in aliprod home, then
they should point to the real folders when the file catalog will reach
its final shape).

What is important is that the new implementation is backward compatible
and the users won't notice the changes, since the flag in the Put methods
is set by default to "Private".

Alberto