Added code to do flow analysis.
The core of the code is in the class AliFMDFlowBin.   It contains an
AliFMDFlowHarmonic object to calculate the harmonic, 3
AliFMDFlowEventPlane objects to calculate the full event event-plane,
and two sub-event event-planes, as well as three AliFMDFlowResolution
objects for calculating the event-plane resolution in 3 different
ways.

The classes AliFMDFlowBinned1D and AliFMDFlowBinned2D are "histograms"
of AliFMDFlowBin objects.

AliFMDFlowAxis is an axis in an AliFMDFlowBinned1D or AliFMDFlowBinned2D
"histogram".

AliFMDFlowTrueBin and AliFMDFlowTrue1D are for calculating the flow
from a known event plane angle - e.g., when reading data from an
event generator.

The class AliFMDFlowStat is single variable statistics that does not
suffer from rounding errors when adding small numbers to a large sum.

AliFMDFlowBessel is  namespace for a number of Bessel functions.  These
allow one to efficiently calculate the value of several Modified Bessel
functions of the 1st kind in one go.

AliFMDFlowUtil contains some utility functions used through out the code.

Example scripts and programs will follow later.