Added shuttle preprocessor code from Hans Dalsgaard (with several corrections).   Added test script script/TestPreprocessor.C too.  The test script generates its own input, runs the (test) shuttle with that input, reads back the newly written values, and compares to the known input.
Added extra convinience member functions to AliFMDAltroMapping, including
the Print member function to dump the map to standard out.

Added parameter to AliFMDParameters::Init to enable getting only a
subset of the calibration parameters.  This is needed by the shuttle
preprocessor in order to be independt of previously written data.