Summary of all changes: - in AliESDCaloCluster the following have been added: UShort_t  fNCells ; UShort_t *fCellsAbsId;   //[fNCells] array of cell absId numbers Double32_t *fCellsAmpFraction;    //[fNCells][0.,1.,16] array with cell amplitudes fraction.
The first 2 data members are not really necessary since the data member  TArrayS * fDigitIndex;
has the same function, but in order to change the name to something that avoids the mention of digits ... and to have the same data member type as fCellsAmpFraction, it was added (could be removed in the future).

- in  AliESDCaloCells we have:
 Int_t       fNCells;       // Number of cells
 Short_t    *fCellNumber;   //[fNCells] array of cell numbers
 Double32_t *fAmplitude;    //[fNCells][0.,0.,16] array with cell amplitudes (= energy!)
 Double32_t *fTime;    //[fNCells][0.,0.,16] array with cell times

The changes in AliESDCaloCluster were propagated to AliAODCaloCluster.

Gustavo