Massive commit to flesh out the geometry in more details. Most of the geometry is there.  Missing - if at all relevant are only
- Cooling pipes in FMD3 
- Cooling fin in FMD2

Although all volumes are there, it should be noted that the 
geometry should really be checked against CAD drawings and
installation/assembly notes.  This will, however, take some 
time to do, and I need some information from the relevant 
parties. 

What has been added to the geometry is 

- Support of FMD2 - barrel, covers, and flanges. 
- Support of FMD1 - both the upper and lower saucers are there.
- Digitiser cards on all half-rings. 
- Cooling plates on FMD3

What has been changed in the geometry is 

- Sensors are now put into an assembly together with the hybrid
  cards and the stand-offs.  This is to make it easier to have 
  enough clearance for alignment, and it is, after all, more likely
  that the mounting of a module (sensor+hybrid+stand-offs) to the 
  honeycomb plates is less precise than the compound itself. 
- The cone of FMD3 is now made as a composite shape of a cone minus 
  the openings where the cooling plates sit. 
- Computations done more robust - gone are a lot of stupid fudge
  parameters.  

Other changes are for the display and alignement macros and similar.