The HLT online display now compiles against standard aliroot installation (No external HLT packages needed).
The script automatically compiles the online display if it is
not allready compiles. It then tries to connect
to the output any PHOS HLT proceesses running on the HLT 
cluster vi port forwarding from
the PHOS nodes to the hlt gateway to Vobox0
wich is visible on the CERN network.