Re-organization of vertex constraints in the primary vertex determination:
- 3 diamonds in OCDB: GPR/MeanVertexSPD for SPD, GRP/MeanVertex for
ITS tracks, GPR/MeanVertexTPC for TPC tracks. The diamond estimated by the 
SPD DA is stored in MeanVertexSPD. This mean vertex can be used to constrain
the SPD vertexer.
- 3 separate flags to have the freedom to use the constraint for
given vertex and not for another. These flags are moved from AliReconstruction
to AliITSRecoParam (for SPD) and AliGRPRecoParam (for tracks), where the other
cuts for the vertexers are. 

Andrea