Package for the SPD  FO-uniformity-scan DA. It is made up of ITSSPDFOda.cxx and the 5 classes, AliITSOnlineSPDfo, AliITSOnlineSPDfoChip, AliITSOnlineSPDfoChipConfig, AliITSOnlineSPDfoInfo and AliITSOnlineSPDfoAnalyzer.
-----------------------------------------------------------------------------
AliITSOnlineSPDfo :
It is a manager class that knows how to store data, how to read them and how to
manage them.

AliITSOnlineSPDfoChip (+AliITSOnlineSPDfoChipConfig):
It is the container class. It stores per each pixel configuration 
inside the chip the corresponding FO output. 

AliITSOnlineSPDfoInfo:
It stores the how the scan was arranged (number of dacs, their codes, number 
of pulses, etc.)

AliITSOnlineSPDfoAnalyzer:
It is the class that analyzes the FO scan data to determine the more suitable 
DAC values to have the best FO trigger efficiency.

Annalisa Mastroserio