Modifications of the activation scheme for module agents and the printout at component library loading
In order to activate only the agents of the library being loaded,
a check for library name and module id is applied. Works for
component libraries with agents following the naming scheme
libAliHLT<MOD>.so

Info messages have been adjusted to the needs in order to reduce
output. The full message is now only printed if the library
contains detailed build information from the HLT build system.