hange of the interfaces.
Now the GlobalVertexer can read tracks directly from the ITS/TPC trackers and can be called before the ESDConverter

The change is backward-compatible, the vertexer can be also called after the ESDConverter, as it was before.
The GlobalVertexerHisto component can read both the ESD and the GlobalVertexer output.
 
The reason for this change is very low speed of the packing/unpacking the AliESDEvent object.
The GlobalVertexer speeded up in >10 times after this change.