This is a backward incompatible change in AliRoot. The following methods have been removed
-   const char*   AliRun:: GetTriggerDescriptor() const {return fTriggerDescriptor.Data();}
-   void                AliRun::SetTriggerDescriptor(const char *name) {fTriggerDescriptor = name;}

as their only function was to store temporarily the beam type in AliRun to be transferred to AliSimulation. This information is used for the trigger configuration. Instead, we have introduced the two methods

+  void                   AliSimulation::SetTriggerConfig(TString conf) {fTriggerConfig=conf;}
+  const Char_t*  AliSimulation::GetTriggerConfig() const {return fTriggerConfig.Data();}

To store and get directly this info from the AliSimulation class. This change is completely transparent, however what is NOT transparent is the setting of the beam type which is done in the Config.C. Typically the change is the following

-    gAlice->SetTriggerDescriptor("Pb-Pb");
+    AliSimulation::Instance()->SetTriggerConfig("Pb-Pb");

All the macros in AliRoot have been changed, but also all private macros will have to be changed. The symptom will be very easy to detect, as CINT will complain that there is no method SetTriggerConfig in AliRun. 

One subtle point here is that with this change Config.C will require the presence of AliSimulation. If it is not there, it will be created by the Instance() call. This is not a problem in normal usage, but if Config.C is not used for simulation, it will create an instance of AliSimulation, that will not be used. Again, this should not be a problem, but users might as well be aware o fit.