Changes to make DAs publish information to AMORE.
The pedestal DA will publish a TH2F for each ring, containg as content
the pedestal values, and as errors the noise values.

The gain DA will publish two TH2F per ring.  One contains the gains 
and the error on the gains, and the other contains the reduced 
chi^2 for each of the strips.  

Both will also publish a TObject with the unique identifier set to the 
run number used to generate the data. 
:
The base DA does not publish anything at this time.