Possibility to (1) Read the vertex position and vertex sigma_z from raw CDB. This overwrites any previous setting in the Config.C
(2) Set the magnetic field from raw GRP entry.
In this case the setting cannot be changed in Config.C

The new GRP object from simulation is written to a location that can be specified
by the user. (CDB is temporarily unlocked for this operation)


These two options move all OCDB access from Config.C to AliSimulation. The usage of CDB entries are
exposed as AliSimulation settings (usually in the sim.C macro used for production).