New implementation of the forward multiplicity analysis.
In this version (unlike the older one), two passes are 
needed:  
  - In the first pass, an AliAODForwardMult object is 
    put in the AOD tree - one per event. 
  - In the second pass, that information is read and 
    the final dN/deta constructed based on the parameters
    selected. 

The scripts RunManager.C and Pass1.C are an example of the
first pass.  

The script DrawRes.C and Pass2.C are examples of the 
second pass.  

Note, both examples are designed to run locally.   

The code is documented using Doxygen (see the ../doc
directory).   A standalone document is also under 
preparation in ../doc/doc.tex.  

The code in 'older' are here for reference and will be removed
shortly.