Update of the TRD PID Response:
- New reference distributions covering a momentum range from 1 to 10 GeV/c with an additional step at 1.5GeV/c. The reference in the OCDB up to now are going up to 6 GeV/c in integer momentum steps. With the change we increase the pt reach for the TRD up to 10GeV/c.

- The container type has changed: Before the references were stored in a TObjArray and associated to momentum steps and species  via the name. Momentum steps were hardcoded in the PID Response, adapted to the old TRD PID code. With the new references having a different momentum binning we have seen that this is difficult to handle. A new class AliTRDPIDReference stores the reference objects together with the reference object container. The new container can handle references for all TRD PID methods. 

- To keep backward compatibility the OCDB object contains the TObjArray containing the old reference objects and the reference container with the new references. In the code currently implemented the Load function checks the name of the TObject, and if it is not valid it will be neglected. In the new code we explicitly expect the new container, and all objects of a different type will be ignored. So backward compatibility is not an issue. One file in the TRD code is affected by the change, which has to be updated the same time  commit is done.

- The TRD tender was also adapted to the change and the configuration was updated.

Concerning the OCDB entry since the TRD PID Response is now used in the reconstruction to calculate the TRD PID Probabilities we were concerned that we would cause problems there if we provide only the corresponding object for the new code while the reconstruction is still running with the old code. For this since the current OCDB entry stores the reference object in a simple TObjArray, we considered to be compatible with both versions of the code we add the new object to the old references. Since in the old code references which are not compatible with the name convention for the old objects are rejected, and in the new code we require the reference object to be of the new type, both versions should work with the new code.

Markus