Updates from David Rohr to add checks for fakes and clones and several bugfixes
Details see below

It follows a list of changes:
- Change AliPerformanceEff::AddHistoEff such that it can create
histograms for secondary tracks.
- Change code for secondary track histogram calculation such that
AddHistoEff is used in order to remove redundant code.
- Fix cuts on particle IDs (secondary plots that were labeled with
electron efficiency contained pion efficiency)
- Change AliPerformanceEff::AddHistoEff such that it can create
histograms for clone and fake rate
- Add clone and fake rate calculation to AliPerformanceEff.
- Add two switches to AddTaskPerformanceTPCdEdxQA, one for using TPC
only tracks, one for creating both primary and secondary track
efficiency plots.
- Fix for AliPerformanceRes: If IsUseTrackVertex() is not set,
vtxESD->GetStatus is not needed, however it will always fail and no
resolution is calculated, call is skipped in this case.
- Print an Error in AliPerformanceRes if tracks cannot be propagated
because magnetic field has not been initialized instead of propagating
with 0-field.
- Always read ESDFriends from esdTree not from HLTEsdTree, otherwise mc
information is not available for HLT tracks.
- Change the track rotation in AliPerformanceRes::ProcessInnerTPC,
rotate the MC parameters not the track parameters, this allows to
propagate track along x where propagation works best
- Fix AliPerformanceRes::ProcessTPC: HLT tracks are not necessarily in
global coordinates but need to be rotated first.
- Fix AliPerformanceRes::ProcessTPC: Originally, tracks were not
propagated to correct MC coordinates.
- Clear the SetUserRanges of the fHistoRes histograms, so future users
do not get confused if they do not know about the cuts.
- Change eta cut in AliPerformanceEff from 0-0.9 to -0.9-0.9
- Do not include fake tracks in AliPerformanceRes resolution calculation
- Ignore all tracks with mc label 0 in both AliPerformanceRes and
AliPerformanceEff
- Use rotated momentum for pullPhi calculation in AliPerformanceRes
(sigma is calculated in terms of Sin(Phi) and GetSnp returns local Sin(Phi))
- The check for primary tracks in AliPerformanceRes uses results from
RelateToVertex function which is not called if IsUseTrackVertex() is not
set, changed such that in this case MC information is querried whether
the track is primary or not.
- Change check whether a track is findable or not as discussed with
Ruben: it is no longer necessary that an ESD track with that label
exists, a findable MC track is sufficient.
- Do not consider tracks with negative label in efficiency as these are
considered fake tracks.
- Fixed calculation of number of daughters inside the Vertex in
AliPerformanceRes, counter must be initialized with 0.
- Reenable AliPerformanceRes in AddTaskPerformanceTPCdEdxQA