Renamed some member functions for more logical names
  Init            -> SetupForData
  DefineOutput    -> CreateOutputObjects
  Output          -> CreateOutputObjects
  InitializeSubs  -> SetupForData
  ScaleHistograms -> Terminate 

since these better reflect when they are executed. 

Output of AliForward{,MC}MultiplicityTask is split into 
two - one for sums and one for Terminate results.  This
is for Grid where the output containers are not opened 
read/write while parameter containers are. 

Added a script that partially automatize the dN/deta 
analysis on the Grid.  A similar script for CAF (or
Proof) should come soon.