Important bug fixes for the PWGLF/FORWARD directory
	modified:   ../../../OADB/PWGLF/FORWARD/CORRECTIONS/data/fmd_corrections.root
	modified:   ../../../OADB/PWGLF/FORWARD/CORRECTIONS/data/spd_corrections.root

Updated corrections and include new noise-gain correction for all possible runs

	modified:   ../../CMakelibPWGLFforward2.pkg
	modified:   ../../PWGLFforward2LinkDef.h

New classes added

	modified:   AddTaskFMDELoss.C

Minor fixes for new AliFMDESDFixer

	modified:   AliAODForwardMult.cxx

Allow trigger mask "CENT"

	modified:   AliBaseESDTask.cxx
	modified:   AliBaseESDTask.h

New method PreCorrections - meant to do stuff when we know the run number etc.,
but before reading the corrections

	new file:   AliFMDCorrNoiseGain.h

A new type of correction.  This is used to add in the additional noise that
the reconstruction failed to add because of a bug in the AliFMDRawReader.  With
this, we can process old data with out fear of corruption.

	modified:   AliFMDDensityCalculator.cxx
	modified:   AliFMDDensityCalculator.h

Moved re-calculation of eta to new AliFMDESDFixer class

	new file:   AliFMDESDFixer.cxx
	new file:   AliFMDESDFixer.h

A class that fixes up the FMD ESD object for various flaws, including
bad values for 0's, missing noise from reconstruction, and possibly
re-calculation of eta.  This class takes an AliESDFMD object and
modifies the content in-place.  Unfortunately we cannot re-calculate
the phis in this, because we never store the phis in the ESD object.

	modified:   AliFMDEnergyFitter.cxx
	modified:   AliFMDEnergyFitter.h

Minor stuff

	modified:   AliFMDEnergyFitterTask.cxx
	modified:   AliFMDEnergyFitterTask.h

Use new class AliESDFMDFixer to fix up ESDs before processing

	modified:   AliFMDEventInspector.cxx
	modified:   AliFMDEventInspector.h
	modified:   AliFMDMCEventInspector.cxx

Added new method for 2013 pA data.  Updated the method for
selecting which vertex method to use.

	modified:   AliFMDSharingFilter.cxx
	modified:   AliFMDSharingFilter.h

Moved ESD corrections into new class AliFMDESDFixer.

	modified:   AliForwardCorrectionManager.cxx
	modified:   AliForwardCorrectionManager.h

Possiblity for new correction object AliFMDCorrNoiseGain.

	modified:   AliForwardMCMultiplicityTask.cxx
	modified:   AliForwardMCMultiplicityTask.h
	modified:   AliForwardMultiplicityBase.cxx
	modified:   AliForwardMultiplicityBase.h
	modified:   AliForwardMultiplicityTask.cxx
	modified:   AliForwardMultiplicityTask.h

Updates for new correction ALiFMDCorrNoiseGain

	modified:   DrawdNdeta.C

Minor things

	modified:   ForwardAODConfig.C

Updates for changes in the code

	new file:   corrs/Browse.C
	modified:   corrs/ForwardOADBGui.C

New file and fixes for browsing corrections database

	modified:   corrs/CorrDrawer.C

Updates for new noise-gain correction

	modified:   corrs/RunTestF.C

Run test of Landau*Gaus shift of Delta_p

	new file:   corrs/Upload.C

Generic script to upload corrections

	modified:   qa/QAPlotter.C
	modified:   qa/QATrender.C
	modified:   qa/RunQAMT.sh
	modified:   qa/runQA.sh

Updates for new centralized QA framework

	modified:   scan/Trend.C
	modified:   trains/MakeFMDELossTrain.C

Minor things.