dN/deta task now configurable through external script.
Better drawing of dN/deta and summary of dN/deta

Added possibility to select centrality estimator to
something stored in the AOD header.  In principle we should do
the same for the IPz - however, one has to be a little careful
since we use the IPz in the ESD->AOD creation.  If one then
chooses another kind of vertex determination in the dN/deta pass
it could cause a bias.  Finally, we should also be able to use
trigger information stored in the AOD header, but I should check
that we do have all the needed information there.