add common abstract interface classes for flat and fat ESDs
once ported, an analysis task (or any user code) will be able to run
online on the HLT and offline on standard ESDs. Porting consists mainly
of moving to the abstract base class interfaces:
  AliESDEvent -> AliVVevent
  AliESDtrack -> AliVVtrack
  AliExternalTrackParam->AliVVtrack
  AliVTrack -> AliVVtrack
  AliVertex->AliVVvertex
  AliKink ->AliVVkink

this way porting should be easy.

things to solve are friend access (to be added soon).

in 2 places: AliESDv0 and AliAODMCParticle we had to change the
signature of GetStatus() since now it is in the base class (AliVVtrack)
- this has to be cross checked with the authors of the code to decide
  what a common return type will be (we have it now ULong_t, hope nobody
wants to return negative values)! we just make it compile now.