add 2 additional nonlinearity corrections to aliroot:
kPi0MCv6 (to be used on MC in conjunction with kSDMv6)
kSDMv6 (to be used on data in conjunction with kPi0MCv6)

These are based on the symmetric decay method but the shapes of the
nonlinearity curves are constrained by the test beam data.  The method
is described in the note: https://aliceinfo.cern.ch/Notes/node/211 -
Sec 3.1.2 (Test Beam Constrained SDM).

Author: Jason Kamin