Bug fix
To save computing time, the trigger classes are not re-computed
for events with the same trigger mask.
However, such events might have different trigger inputs.
This was not taken into account, thus providing a wrong response
for two sequential events with exactly the same trigger mask but a different input
(when requiring a pattern with the input).
Such cases are rare: yet, this was a bug. Now fixed.