Implementation of fiducial area cut
Implemetation of fiducial area cut. The fiductial area cut is combined
with the number of crossed rows cut (for AODs) in the class
AliEMCalTriggerExtraCuts. The virtual track selection classes are
changed in the way that they can handle multiple cut objects. In the
analysis task itself, all tracks are checked whether the pointer back
to the original event is set, and if not, the pointer is set in that
place. The fiducial area cut is added to the standard track selection
cuts in the add macro of the task.