Implementation of a reduced tree structure for high-pt track anaylses
The reduced event structure contains filtered information about
- Reconstructed event
- Monte-Carlo specific quantities
- True particles
- Reconstructed EMCAL trigger patches
- Reconstructed EMCAL clusters
- Reconstructed tracks
Tracks can be linked to a cluster and a particle via the index. For tracks several selections are applied at filter level. A bitmap stores the indices of the cuts under which the track was selected. A producer task, inheriting from AliAnalysisTaskEmcal, creates the reduced event by doing the filtering and writes to the tree. Reduced trees can be created both from ESDs and AODs, if a track selection for AODs is implemented.