Add histograms for tracks matched to EMCAL trigger patches
Histograms for tracks where reconstructed tracks are matched to EMCAL
trigger patches are added to the tracks analysis component. Matching is
defined as
- Tracks have to be within the boundaries of the EMCAL patches at the
  EMCAL surface
- The patch type has to be the same as the event trigger type
The matching is done either to online or offline patches, but not both
at the same time. In order to get the information which patch type is
used, a getter for the trigger decision configuration is added to the
trigger decision.