first version of the ZMQ sink/publisher component
the component can operate as a server (REQ/REP) or as a PUBlisher.
in REP mode the REQ contains a string (may be empty) to select the data
type to be sent back. e.g. for ESDs it will be: ALIESDV0, for flat ESDs: FLATESD0
(strings should not be null terminated).