Handle trigger selection using trigger class manager
New approach to handle trigger selection: A trigger class manager
contains a list of trigger classes (can be defined per period) and
selects the matching ones per event. Furthermore it steers the
histograms per trigger class. A trigger class, defined in the add macro,
consists of a name, a title and a method to select triggered events,
which can be either a trigger bit, a pattern (can consist of several
parts) in the trigger string or a set of trigger patches found in the
EMCAL. The trigger selection is done in the event loop inside the
analysis task, and the trigger manager is passed to the single analysis
components. In this case  it is possible to implement different trigger
selections for different periods. For pp 2012 and pPb 2013 the trigger
setup is implemented in the add macro.