Add virtual track selection to the EMCAL framework
Add classes for virtual track selection to the EMCAL framework. These
classes are derived from AliEMCalPtTaskVTrackSelection, with the following
differences:
- New classes are not part of a specific namespace
- New classes allow to handle conditions "AND" and "OR" of track cuts
  differently
- New classes internally work completely with AliVCuts, a downcast is
  only necessary when applying AliESDtrackCuts to AODs in order to
  call the copy constructor of AliESDtrack
These classes allow track selection for ESDs and AODs in a transparent
way so that the user tasks can stick to the base classes (AliV*) and
consequently run natively on ESDs and AODs without modification. AliPicoTracks
are suppported as well, but the support will be removed when the pico
tracks become abandoned.