Fix bug limiting area for jet patch algorithm
In order to circumvent a problem that patches overlapping the last
two supermodules, which were inactive, were found by the reconstruction,
a (temporary) fix was addd to the trigger maker limiting the active
area to the 10 big supermodules. The cut was now hard coded in the trigger
maker to be insensitive to changes in the geometry.