Initial alihadd work, see also ALIROOT-5727
This introduces a simple utility, alihadd, whose final goal is to extend
ROOT hadd to match the usecases presented in ALIROOT-5727.

For the moment it should simply provide a drop in replacement for hadd
with the additional ability to merge only the TKeys which are specified
via (eventually multiple) the -i option. E.g.:

  alihadd -i "h1.*" -- output.root input1.root input2.root

should create an output.root file with all the tkeys which start with
"h1".