Prepare EMCAL trigger maker for handling of DCAL triggers
In order to be able to handle patches from DCAL as well,
the data container with ADC and amplitudes need to be able
to handle variable size depending on whether DCAL was available
or not. Trigger data are no longer cached in a hard coded
array but in a container with flexible size, called
AliEmcalTriggerDataGrid. In order to allow for different data
content, the new class is made a template class which supports
primitive number data types. The container is allocated in
ExecOnce. In contrast to the previous implementation the data
grid is made boundary-safe throwing an OutOfBoundsException
when inexisting rows or cols are accessed. Furthermore it can
be dynamically increased, decreasing is not possible.