Updates for GCC 5 and ROOT 5.34/35
Protected sub-classes cannot be accessed from dictionary code.
For the most part, we no longer make dictionaries for these
kinds of classes.  There's a few notable exceptions - places
where objects of the sub-class are actually streamed - e.g., as an
analysis train is transmitted to Grid or Proof worker nodes via
a file.