Improve AliTHn to allow to use also double precision in the storage containers: - AliTHn uses float as usual - AliTHnD uses double (this needs twice the memory, so only use it if there is a reason for double precision).
Further, there is a common base class AliTHnBase for the two.

This change is fully backward compatible, i.e. files produced before this change can be read with the updated AliPhysics.

However, it is not forward compatible. You cannot read with an old AliPhysics, output produced with AliPhysics versions starting from today. So if you are an AliTHn user, please update your AliPhysics (to the master or to the tag which will appear later today).