This fix is no longer needed, now that AliDAQ.cxx is fixed
Revert "fix 'invalid detector ID' issue in GlobalTrigger. AliDAQ seems to use (kNDetectors - 1) internally for the HLTId, but it uses kHLTId = 30 for determining the HLT DDLS. AliHLTDAQ Uses kHLTId always, so there needs to be some conversion. Perhaps this should be cleaned up and the HLTID should be made consistent in the future"

This reverts commit f740663ac22517c27f6d329ad03813e6141712de.