Split specific analysers into files.
Defined factory of analysers.
Factory has makers of analysers.

Each script should define it's maker class.
Makers can be instantized (and registered) in the ProcessFast.C script.
Each maker registers it self with the factory under a name (type)
Different sub-types of analysers can be made by the maker.

This makes it very easy to add new analysers as one sees fit.

To do this, one specifies the "type" option in the execution URL.
The factory then matches this to the declared name of the maker.
The factory then passes the sub-type to the maker, which should
then make an analyser (or return null in case it cannot).

Each maker should register it's script name so that it can be
uploaded to PROOF if needed.

Users can define their own analysers and corresponding maker
in a separate script, and copy the ProcessFast.C script to
instantize the maker.  E.g.,

  struct FooBaseAnalyser : public FastAnalyser
  {
    ...
  };
  struct Foo1Analyser : public FooBaseAnalyser
  {
    ...
  };
  struct BarAnalyser : public FastAnalyser
  {
    ...
  }
  struct FooBarMaker : public FastAnalyser::Maker
  {
    FastAnalyser* Make(const char* sub, ...)
    {
      TString s(sub);
      if (sub.EqualTo("foo1",TString::kIgnoreCase))
        return new Foo1Analyser(...);
      else if (sub.EqualTo("bar",TString::kIgnoreCase))
        return new BarAnalyser(...)
      return 0;
    }
    const char* Script() const { return __FILE__; }
    void List() const
    {
      Printf(" Foo1");
      Printf(" Bar");
    }
  };