Make jet patch size adaptable
As in PbPb 2015 data is taken with a jet patch size of 16, the
jet patch size is made adaptable. Consequently patch types are
no longer identified by the size of the patch but by the bitmask
in the trigger algorithm, which is defined when the patch finders
are created. In case the jet patch size is 8, the background
patch finder will not be created, but instead the bitmask of the
jet patch finder will also contain the bit of background patches