Fix the DIPSY issue with track->Pt() == 0
It was found that dipsy produces primary particles (mainly protons) with
0 momentum. This can cause FPE down the road. Previously, this problem
was tackled by checking if the pdg code of each track seems
valid. However, since even protons may have pT = 0, this failed to
mitigate the problem. This commit reverses this previously implemented
check and instead checks for Pt() > 0 if necessary for the given observable

This reverts commit 2d9575324b29bbcfb08b406d0e480a5ff17e8257.