PWGCF: AliFemtoEventReaderAOD - Fixed major performance bug.
The bug was in CopyPIDtoFemtoTrack method: for every track in an event, this method would call GetNContributors which is
not constant in time, but linear with the number of tracks in event, essentially creating an n^2 time complexity.

The solution was to employ a static std::map as a cache to save the result of GetNContributors to the unique id of the
vertex. This is checked first and will use the cached value. This has shown tremendous speedup - reducing runtimes by
over 70%.