FIX: use cpp casts for well defined behavior
The bug was as follows: Q2 (aka GetpTHard() was a memeber of the pythia
header. The cast of this header was done with the old C way and did not
fail for Dipsy. The C way means that no type-checking is done at runtime
and it might even resort to reinterpreted cast. Ie, the dipsy data in memory
was used "as is" through the pythia structure. This is clearly undefined
behavior and its strange that it did not crash sooner than it did.

The fix is to use dynamic_cast instead to check if the header in memory is truly a
pythia one. Only then are the member functions for nMPI and Q2 called!