Bug fix for Proof case.
This task was checked only locally on a TChain, where the AliAODEvent object is the same throughout the chain (i.e. it survives each file opening of the chain). That’s not the case in a proof scenario, where the AliAODEvent is deleted (and so is the fDimuons object attached to it…) re-created at each file.