Fix vertex selection in AliAnalysisTaskEmcal and doxygen
Vertex selection was previously NOT handled by the
AliAnalysisTaskEmcal in case the AliAnalysisUtil was
used in addition due to improper handling of the vertex
upper limit. Fixed in this commit. It has been taken
care of that user-defined vertex selections are not
overwritten. In addition some some doxygen documentation
is added to the AliAnalysisTaskEmcal.