Implementation of stl-iterators in EMCAL container classes
Two types of stl iterators are implemented for classes
inheriting from AliEmcalContainers
- Iterator over all objects
- Iterator over accepted objects
As iterators must support also const containers, the
object to iterate over is always a const container,
therefore all simple accessor functions are made
constant.

A unit test is implemented for the iterators of the
classes AliClusterContainer and AliParticleContainer.