New tracklet code which works in two steps - First and ESD->(custom) AOD pass.   In this pass, we re-run the   reconstruction, and does the injection of clusters.  The resulting   tracklets are stored in a separate branch on the AOD tree.   For simulated data, we also extract the generated "tracklet"   distribution and flag tracks as one of from a primary or secondary   particle, or from combinatorial background.  Note, no event selection   is imposed in this pass. - Second AOD->partial result pass.  In this pass we impose a   specific event (and possibly centrality) selection.  We histogram   tracklet variables (eta, delta) in specific histograms based on the   type of tracklet (measured, from injection, from combinatorics,   directly from generator).
These two passes should be done for both real and simulated data.

The ideas is, that we do the first ESD->AOD pass once.  We then have
the tracklets and events that we can impose different cuts on.  In
this way, cut variations and other manipulations (e.g., simulation
weights) can be applied at the AOD level which allows for much faster
turn-around time then the ESD analysis.

Yet to be implemented is the code to take the real and simulated
data together to make the final results.  However, most of that
code exist already in ../tracklets2, so it should be easy to do.