c++11 range-based iteration compatibility of EMCAL containers
In order to provide c++11 range-based iterator compatibility for
both all and accepted objects inside the EMCAL containers, an
interface AliEmcalIterableContainer including its own stl
iterator is created. This interface handles both accepted and
all objects in the container, depending on the specification
in the constructor.

Iterating over objects inside the Cluster-/Particle-/Track-/Jet-
container in case of c++11 support reduces now to simply
- for all objects:
  for(auto en : cont->all()) {...}
- for accepted objects:
  for(auto en : cont->accepted()) {...}
As the iterator interface is implemented for the base class
AliEmcalContainer (and consequently all deriving classes supported),
a downcast from en to the corresponding type of object inside the
container is required.

Iteration can be performed on constant containers.