Optional event selection according to MultSelection
Now it is optional to use the same event selection than the
one used for calibrating centrality in MultSelection apart
from the own task centrality and zvertex cuts.

With the use option in yes we get valid centrality (multiplicity)
percentile only for the events used for calibrating centrality
and the reject reason (wrong percentile) for the ones not
accepted.