Implement method for automatic downscale correction
For periods with calo triggers the min. bias triggers, L0
triggers and (if available) low threshold L1 triggers are
randomly downscaled by a known factor. The downscaling
factors are known, but changing run-by-run. In order to
handle the run dependence the downscaling factors are read
from an OCDB container. They are applied as weight using
1/downscale for all histograms connected to the given
trigger class.

Normally high-threshold histograms are a subset of the low-
threshold and non-threshold triggers. Downscaling swaps this
condition: A subset of the low-thershold triggers are a subset
of the high-threshold triggers. In order to make the samples
independent the low threshold triggers need to be remove from
the high-threshold triggers.