Commit of yaml-cpp as a static lib
Added yaml-cpp to PWG/Tools/yaml-cpp so that it is available to all
users.

Commited version 7d2873ce9f2202ea21b6a8c5ecbc9fe38032c229, "Fix scalar
parsing when a line starts with a comment.". A tagged version would
have been used, but the switch to c++11 was made after the most recent
tag (boost was previously required). This should be updated to the
next tag when it is available!

The only modifications to the source included removing the "test"
directory, which added a substantial amount of code and compile
time for little benefit for our use (we aren't running the test). The
test module was explicitly disabled in PWG/Tools/CMakeLists.txt.