Protection for file handles using C++11 std::unique_ptr
File handles put inside std::unique_ptr. The unique_ptr will
delete the file, and consequently close it, for sure once
the pointer gets out of scope. In this case the files will be
closed even a new condition is added or an exception is raised.
Also the result of TString::Tokenize is handled with
std::unique_ptr.