Adapt handling of the offline and recalc patches in the old trigger maker to the patch structure of the new trigger maker
Handling of the offline and recalc trigger bits are different in the old and the new trigger maker.
The old trigger maker was mixing the offline and recalc bits with online bits, as it was done
formerly. Consequently offline and recalc patches are not identified as by now separate bit ranges
are used for online, offline and recalc patches.

The algorithm is modified in the way that in the simple offline trigger online and offline bits will
be set in parallel once the patch is accepted (cutting on the energy threshold). Inside ProcessPatch,
which calculates then patch kinematics, online bits are removed.

In addition, in order to avoid double counting, the maximum patch is no longer added to the offline
EMCAL triggers as the patch will appear a second time as regular patch. The indices for the maximum
patch will only be used for debugging / QA.

With these changes the old trigger maker can be kept alive for a transition period, usage is however
not recommended as the new trigger maker is more flexible, robust and efficient.