Bugfix in loading pythia header
The Pythia Header is obtained in ExecOnce for the
first event. In this case it doesn't change any more.
The pt-hard value obtained in this case is unreliable
and the outlier rejection is not supposed to work
properly.

With this fix the Pythia Header is obtained in
RetrieveEventObjects. In this case the Pythia Header
is obtained for each event.