Implement control probes in the trigger mimic task
Adding histos for the control probes EMCAL clusters (AliAnalysisTaskEmcalClustersRef)
and offline patches (AliAnalysisTaskEmcalPatchesRef) to the trigger mimic task and
handle the same way as reconstructed tracks (no cut on a trigger class name, instead
tunable threshold in the event selection).

Setters are added for different observables. This is needed in order to switch on/off
components in case the LEGO trains become too heavy.