EMCal Embed: Core implementation
The new EMCal Embedding framework enables uses to embed various MC
signals into already reconstructed events to understand the jet energy
scale. The framework is discussed at:
https://alice.its.cern.ch/jira/browse/ALPHY-53

The general approach is to create additional AliEmcalContainer derived
classes which contain the embedded signals. Then, each task will loop
over the available containers, thereby embedding the signals. This has
a number of benefits:
- Support of looping over multiple containers is useful beyond
  embedding.
- Embedding comes "for free". It "just works".
- Embedding specific code is encapsulated, such that tasks can use it
  without knowing about the embedding

Some features include:
- An option to apply the track efficiency only to the embedded
  container instead of to all containers. This is implemented in the jet
  finder
- Basic embedded event selection. More sophisticated event selection
  will have to be performed outside of the main embedding task, as
  objects such as jets are not yet available.
- Steering of the EMCal Corrections is handled by the EMCal Correction
  Framework and comes for free.

Internally, we create an additional event, referred to as an "external
event" which contains the embedded AOD or ESD tree. Once this event is
passed to the AliEmcalContainer derived class, everything works as
normal, but using the embedded objects. To support this, a TChain is setup
to handle the input so the Ali{AOD,ESD}Event itself never changes - only
the underlying tree does. This requires manual management of the
location within a tree to ensure all desired features are supported
(such as random access into a tree without jumping between many files,
which would stress an AliEn connection). This approach is used because
using TTrees and manually loading the files was not compatible with the
configuration of AliEmcalContainers.

All of this machinery is managed by AliAnalysisTaskEmcalEmbeddingHelper,
which is a singleton class.