Implement online FastOR masking in the trigger maker
Masked FastORs currently need to be derived from the
OCDB. When defined, FastOR ADC values are automatically
neglected for fastors which are masked. It can be
specified that cell energy values from cells contributing
to a masked fastor are neglected as well - in this case
both online and offline patches share the same acceptance.