Fix handling of TRU reading
In run1 the TRUs in the small supermodules were counted in the
number of TRUs in the geometry, however they were inactice and
consequently didn't leave a signal in the preprocessor. This
produces a crash when looping over the TRUs as well. Instead the
loop limits are taken now as well from the DCS config.