Fix mapping between mask and channel for run2 data
The mapping between channel and mask position is linear in run1,
but more complicated in run2 as the TRUs are designed in subregions.
A closure which handles properly the mapping for the two cases. In
case the parameters are out of range an exception is issued which is
handled in a separate error handler branch.